# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/leads": {
      "get": {
        "operationId": "LeadsController_list",
        "parameters": [
          {
            "name": "band",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "immediate",
                "high",
                "investigate",
                "ignore"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            }
          },
          {
            "name": "signalType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "F-LEG",
                "F-PLAT",
                "S1",
                "S2",
                "S3",
                "S4",
                "S5",
                "S6",
                "S7"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "score",
              "type": "string",
              "enum": [
                "score",
                "recency"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "leads"
        ]
      }
    },
    "/api/leads/{id}": {
      "get": {
        "operationId": "LeadsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          }
        },
        "tags": [
          "leads"
        ]
      }
    },
    "/api/leads/{id}/decision": {
      "post": {
        "operationId": "LeadsController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "leads"
        ]
      }
    },
...
    "/api/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            }
          }
        },
        "tags": [
          "companies"
        ]
      }
    },
    "/api/contacts": {
      "get": {
        "operationId": "ContactsController_list",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "contacts"
        ]
      },
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        },
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/scoring-config": {
      "get": {
        "operationId": "ScoringConfigController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringConfigResponse"
                }
              }
            }
          }
        },
        "tags": [
          "scoring-config"
        ]
      },
      "patch": {
        "operationId": "ScoringConfigController_patch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoringConfigPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoringConfigResponse"
                }
              }
            }
          }
        },
        "tags": [
          "scoring-config"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rubico Lead Engine API",
    "description": "Consumed by the React dashboard SPA, served same-origin behind a proxy. Authenticated by an httpOnly session cookie (§8.3).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rle_session"
      },
      "sessionBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CompanyDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "canonicalDomain": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "nullable": true,
            "type": "string"
          },
          "region": {
            "nullable": true,
            "type": "string"
          },
          "headcountBand": {
            "nullable": true,
            "type": "string"
          },
          "industry": {
            "nullable": true,
            "type": "string"
          },
          "detectedStack": {
            "nullable": true
          },
          "legacyFlags": {
            "nullable": true
          },
          "atsProvider": {
            "nullable": true,
            "type": "string"
          },
          "atsSlug": {
            "nullable": true,
            "type": "string"
          },
          "suppressionReason": {
            "nullable": true,
            "type": "string",
            "enum": [
              "client",
              "competitor",
              "rejected",
              "do-not-contact"
            ]
          },
          "firstSeenAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "lastEnrichedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "signals": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "signalId": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "F-LEG",
                    "F-PLAT",
                    "S1",
                    "S2",
                    "S3",
                    "S4",
                    "S5",
                    "S6",
                    "S7"
                  ]
                },
                "eventDate": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "sourceName": {
                  "type": "string"
                },
                "sourceUrl": {
                  "type": "string"
                },
                "excerpt": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "signalId",
                "type",
                "eventDate",
                "sourceName",
                "sourceUrl",
                "excerpt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "canonicalDomain",
          "name",
          "country",
          "region",
          "headcountBand",
          "industry",
          "detectedStack",
          "legacyFlags",
          "atsProvider",
          "atsSlug",
          "suppressionReason",
          "firstSeenAt",
          "lastEnrichedAt",
          "signals"
        ],
        "additionalProperties": false
      },
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "nullable": true,
            "type": "string"
          },
          "seniority": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "emailStatus": {
            "nullable": true,
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "creditCost": {
            "type": "number"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "companyId",
          "name",
          "role",
          "seniority",
          "email",
          "emailStatus",
          "source",
          "creditCost",
          "resolvedAt"
        ],
        "additionalProperties": false
      },
      "ContactListResponse": {
        "type": "object",
        "properties": {
          "contacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "companyId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "role": {
                  "nullable": true,
                  "type": "string"
                },
                "seniority": {
                  "nullable": true,
                  "type": "string"
                },
                "email": {
                  "nullable": true,
                  "type": "string"
                },
                "emailStatus": {
                  "nullable": true,
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "creditCost": {
                  "type": "number"
                },
                "resolvedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "companyId",
                "name",
                "role",
                "seniority",
                "email",
                "emailStatus",
                "source",
                "creditCost",
                "resolvedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "contacts"
        ],
        "additionalProperties": false
      },
      "CreateContactBody": {
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "role": {
            "type": "string",
            "maxLength": 200
          },
          "seniority": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "leadId",
          "name"
        ]
      },
      "DecisionBody": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "reasonCode": {
            "type": "string",
            "enum": [
              "wrong_fit",
              "stale",
              "already_known",
              "no_real_need",
              "bad_contact",
              "good",
              "wrong_archetype",
              "no_technology_need",
              "evidence_too_weak",
              "wrong_capability",
              "platform_not_problem",
              "attribution_error"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "decision",
          "reasonCode"
        ]
      },
      "DecisionResponse": {
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "scoreAtDecision": {
            "type": "number"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "leadId",
          "status",
          "scoreAtDecision",
          "decidedAt"
        ],
        "additionalProperties": false
      },
...
      "LeadListResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "leads": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "companyId": {
                  "type": "string"
                },
                "canonicalDomain": {
                  "type": "string"
                },
                "companyName": {
                  "type": "string"
                },
                "fitScore": {
                  "type": "number"
                },
                "intentScore": {
                  "type": "number"
                },
                "compoundBonus": {
                  "type": "number"
                },
                "totalScore": {
                  "type": "number"
                },
                "band": {
                  "type": "string",
                  "enum": [
                    "immediate",
                    "high",
                    "investigate",
                    "ignore"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "new",
                    "approved",
                    "rejected"
                  ]
                },
                "likelyNeed": {
                  "nullable": true,
                  "type": "string"
                },
                "archetype": {
                  "nullable": true,
                  "type": "string"
                },
                "opportunityKey": {
                  "type": "string"
                },
                "confidence": {
                  "nullable": true,
                  "type": "string"
                },
                "scoredAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "companyId",
                "canonicalDomain",
                "companyName",
                "fitScore",
                "intentScore",
                "compoundBonus",
                "totalScore",
                "band",
                "status",
                "likelyNeed",
                "archetype",
                "opportunityKey",
                "confidence",
                "scoredAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "leads"
        ],
        "additionalProperties": false
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "email",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
      "ScoringConfigPatch": {
        "type": "object",
        "properties": {
          "updates": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                "value": {
                  "type": "number"
                }
              },
              "required": [
                "key",
                "value"
              ]
            }
          }
        },
        "required": [
          "updates"
        ]
      },
      "ScoringConfigResponse": {
        "type": "object",
        "properties": {
          "config": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "value": {
                  "type": "number"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                },
                "updatedBy": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "key",
                "value",
                "updatedAt",
                "updatedBy"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "config"
        ],
        "additionalProperties": false
      },
...
    }
  }
}